优化代码

+++ Java/spring-boot/16.spring boot sso by auth2.0 optimization/spring-boot-sso-client1/src/main/java/com/glitter/spring/boot/web/interceptor/LoginInterceptor.java
@@ -45,7 +45,6 @@ public class LoginInterceptor implements HandlerInterceptor {
     @Autowired
     private IOauthClientInfoService oauthClientInfoService;
 
-
     @Override
     public boolean preHandle(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse, Object handler) throws Exception {
         String loginUrl = GlitterConstants.DOMAIN_SSO_CLIETN1 +"/login";
@@ -68,12 +67,6 @@ public class LoginInterceptor implements HandlerInterceptor {
             return false;
         }
 
-        String jsessionIdEffective = commonCache.get(cacheKeyManager.getLimitMultiLoginKey(String.valueOf(userInfo.getUserId())));
-        if (StringUtils.isBlank(jsessionIdEffective)) {
-            this.response(httpServletRequest, httpServletResponse, loginUrl, "-2", "未登录或会话超时，请重新登陆。");
-            return false;
-        }
-
         // 登录状态jsessionIdCookie会话维护了一个accessToken与全局会话保持通讯
         OauthAccessToken oauthAccessTokenDb = oauthAccessTokenService.getOauthAccessTokenByJsessionid(jsessionIdCookie);
         if (oauthAccessTokenDb == null || StringUtils.isBlank(oauthAccessTokenDb.getAccessToken())) {
@@ -81,6 +74,12 @@ public class LoginInterceptor implements HandlerInterceptor {
             return false;
         }
 
+        // 校验单端登录
+        String jsessionIdEffective = commonCache.get(cacheKeyManager.getLimitMultiLoginKey(String.valueOf(userInfo.getUserId())));
+        if (StringUtils.isBlank(jsessionIdEffective)) {
+            this.response(httpServletRequest, httpServletResponse, loginUrl, "-2", "未登录或会话超时，请重新登陆。");
+            return false;
+        }
         if (!jsessionIdCookie.equals(jsessionIdEffective)) {
             // 被其他端"挤掉"了,注销局部会话。
             oauthAccessTokenService.deleteByJsessionid(session.getId());
@@ -93,90 +92,141 @@ public class LoginInterceptor implements HandlerInterceptor {
             return false;
         }
 
-        // 校验access_token是否有效，只要返回值不是成功0,就认为是失效如果已经过期，或者有效期剩余小于30秒，则认为已过期，则触发refresh_token换access_token过程。
-        // 如果refresh_token换access_token失败，一般情况是refresh_token也失效，则触发注销局部会话，并重定向到sso登录页重新走登录流程。
-        // 因为不注销也不行了。不注销后续没有办法与sso用户中心通讯了，比如续期等操作都没办法做。
-        // 如果sso全局会话还存在的话,这个过程对用户是不需要输入用户名密码的（但是用户会感觉明明点一个操作到某页，但怎么突然到首页了，哈哈）,这根绳上换新的access_token和jsessionid_client这两个蚂蚱。
-        // 如果sso全局会话不存在的话,则当前客户端为全新sso整体会话建立的第一个客户端,在全新的这个整体会话中,维系了一根新绳子和新蚂蚱。
-
-        // 调用auth后，即使sso全局会话失效(比如操作其他客户单退出登录操作了)
-        // 那么该客户端a也会自动触发重新登录,用户名输入用户名和密码后,会重新建立并加入一个全新的全局会话B，
-        // 其他客户端b也是如此处理,则其他客户端执行到此处时,也会自动加入新的全局会话B中,并且此时用户也是免登录的。
-        // 有了这个验证，其实sso注销全局会话时，只注销自己就行，不回调客户端的注销。这样客户端走到此验证时，发现验证不过，都会自行加入新的全局会话。
-        // 但是给用户的体验不太好，首先用户点击a客户端某资源连接时，会收到提示【系统异常，请重试】，当用户重试时，会调转到登录页，重新做sso全局登录，
-        // 当用户点击b客户端某资源时，也会提示【系统异常，请重试】，当用户重试时，会跳转到b的首页，而不是用户想访问的那个资源(跟a客户端一样,重新加入了sso新会话，由于a时已经完成了登录,b是免登的)
-        // 这么做从安全性上没问题，也是完成了sso的功能，但用户体验实在是不好，所以我们还是注销全局会话时，注销调所有局部会话，如果有问题及时给出预警，有问题排查问题，最终使系统达到一个稳定的状态。
-        // 相当于一个是拉的模式，一个是推的模式，我们采用推的模式。
-
-        ResponseResult<Map> responseResult = null;
-        Map<String, String> map = null;
-        Long remainingExpirationTimeLong = null;
+        // 客户端登录校验通过的情况下,需要验证accessToken是否有效,目的有二:
+        // 1.当前accessToken如果无效,需要及时换取新的accessToken,以保持客户端和sso认证中心之间通信的畅通。
+        // 2.accessToken其实代表了sso全局会话,因为是一根绳,通过accessToken可以窥探全局会话是否失效,如果accessToken对应的sso全局会话失效,那么就要注销本地会话,走重新登录流程。
         try {
-            responseResult = ssoRemote.auth(oauthAccessTokenDb.getAccessToken());
-            map = responseResult.getData();
-            if (responseResult.getCode().equals("0")) {
-                // TODO
-            } else {
-                // TODO 过期的情况,需要换accessToken。
-            }
-
-            // accessToken有效的情况下取出remaining_expiration_time
-            Object remainingExpirationTimeObj = map == null ? null : map.get("remaining_expiration_time");
-            String remainingExpirationTimeStr = remainingExpirationTimeObj == null ? null : String.valueOf(remainingExpirationTimeObj);
-            remainingExpirationTimeLong = remainingExpirationTimeStr == null ? null : Long.valueOf(remainingExpirationTimeStr);
-            // 如果accessToken有效的情况下取出remaining_expiration_time为空,则是系统对接的异常情况,要做邮件预警,并线下反馈,排查解决sso_server项目的问题。
-            if (null == map || null == remainingExpirationTimeLong) {
-                throw new BusinessException(CoreConstants.REQUEST_PROGRAM_ERROR_CODE, "系统调用异常");
-            }
+            this.authAccessToken(oauthAccessTokenDb.getAccessToken());
         } catch (Exception e) {
-            e.printStackTrace();
+            // 注销本地会话,走重新登录流程,走重新登录流程有两种情况
+            // (为什么sso全局会话可能会是失效呢,因为用户可能先通过其他客户端点击退出登录操作了,此时全局会话已经失效,然后再切换到当前客户端页面触发资源请求操作,即本次代码流程)
+            // 1.sso全局会话有效,则重新登录流程不需要用户参与,后台默默走->code->accessToken->userInfo->局部会话的流程,然后跳转到系统首页。这根绳上换新的access_token和jsessionid_client这两个蚂蚱。
+            // 2.sso全局会话失效,则重新登录流程需要用户输入用户名和密码,然后走->code->accessToken->userInfo->局部会话的流程,然后跳转到系统首页。这是一个全新的sso全局会话整体,绳和蚂蚱都是新的。
             oauthAccessTokenService.deleteByJsessionid(session.getId());
             session.invalidate();
-            this.response(httpServletRequest, httpServletResponse, loginUrl, "-2", "系统异常,请重试");
+            // 如果是ajax请求,当前端收到返回码是-3时,则直接跳转到loginUrl,不再进行弹窗提示,这是最友好的。
+            // 如果非ajax请求,则也是直接跳转到loginUrl,不再进行弹窗提示,这是最友好的。
+            this.response(httpServletRequest, httpServletResponse, loginUrl, "-3", "系统异常,请重试");
             return false;
         }
 
-        // 如果accessToken未过期但是
-        if (remainingExpirationTimeLong < 30L ) {
-            // 使用refreshToken换accessToken
-            OauthClientInfo oauthClientInfo = oauthClientInfoService.getOauthClientInfoByServerType("sso");
-            String client_id = oauthClientInfo.getClientId();
-            try {
-                Map refreshTokenMap = ssoRemote.refreshToken(client_id, oauthAccessTokenDb.getRefreshToken(), "refresh_token");
-                String access_token = null == refreshTokenMap ? null :(String)refreshTokenMap.get("access_token");
-                String token_type = null == refreshTokenMap ? null : (String)refreshTokenMap.get("token_type");
-                String expires_in = null == refreshTokenMap ? null : (String)refreshTokenMap.get("expires_in");
-                String refresh_token = null == refreshTokenMap ? null : (String)refreshTokenMap.get("refresh_token");
-                String scope = null == refreshTokenMap ? null : (String)refreshTokenMap.get("scope");
-                Long userId = null == refreshTokenMap ? null : Long.valueOf((String)map.get("userId"));
-                // 其他字段这里不一一判断了
-                if (null == refreshTokenMap || StringUtils.isBlank(access_token)) {
-                    oauthAccessTokenService.deleteByJsessionid(session.getId());
-                    session.invalidate();
-                    this.response(httpServletRequest, httpServletResponse, loginUrl, "-2", "系统异常,请重试");
-                    return false;
-                }
-                // 更新accessToken信息
-                OauthAccessToken oauthAccessToken = new OauthAccessToken();
-                oauthAccessToken.setAccessToken(access_token);
-                oauthAccessToken.setExpireIn(StringUtils.isBlank(expires_in) ? null : Long.valueOf(expires_in));
-                oauthAccessToken.setRefreshToken(refresh_token);
-                oauthAccessTokenService.modifyById(oauthAccessToken);
-            } catch (Exception e) {
-                e.printStackTrace();
-                oauthAccessTokenService.deleteByJsessionid(session.getId());
-                session.invalidate();
-                this.response(httpServletRequest, httpServletResponse, loginUrl, "-2", "系统异常,请重试");
-                return false;
-            }
-        }
-
-        // sso全局会话续期
+        // 不管accessToken本来就是有效的,还是通过refreshToken换取了最新的accessToken,代码运行到此处都能保证当前的库中的accessToken是有效的,sso全局会话这一根绳是串起来的,sso全局整体会话是有效的。
+        // 由于上述流程已经为客户端局部会话完成了续期,现在对sso全局会话续期。
         OauthAccessToken oauthAccessTokenDb2 = oauthAccessTokenService.getOauthAccessTokenByJsessionid(jsessionIdCookie);
         ssoRemote.keepAlive(oauthAccessTokenDb2.getAccessToken());
 
         logger.info("LoginInterceptor.preHandle验证成功,jsessionId:{},userId:{},fullName:{}", jsessionIdCookie, userInfo.getUserId(), userInfo.getNickName());
         return true;
+    }
+
+    /**
+     * 校验accessToken有效性
+     * 如果有效则罢了
+     * 如果无效{
+     *     则触发refreshToken换取accessToken,最终得到一个新的有效的aceessToken,维持一根绳
+     *     {
+     *         如果这个过程顺利则罢了
+     *         如果这个过程失败或者异常,则抛出异常,由调用方注销局部会话,并且走重新登录流程.
+     *
+     *         特注:一般情况是refresh_token也失效,导致其换取access_token失败,不管什么原因导致换取失败,此时都要抛出异常,由调用方注销局部会话,并且走重新登录流程.
+     *              因为不注销也不行了,这根绳已经断了,不注销后续没有办法与sso用户中心通讯了,比如续期等操作都没办法做,只能注销通过重新登录重新把这根绳建立起来。
+     *     }
+     * }
+     * 如果accessToken对应的sso全局会话已经失效,sso全局会话整体的前提已失,accessToken没有意义,局部会话没有意义,一根线已经断开无法维系。此时当则抛出异常,由调用方注销局部会话,并且走重新登录流程.
+     * 如果是其他异常情况,一般是sso_server系统报错。这证明当前这跟线不稳定,sso全局会话整体的前提已失,accessToken没有意义,局部会话没有意义,一根线已经断开无法维系。则抛出异常,由调用方注销局部会话,并且走重新登录流程.
+     *
+     * 这个方法本身也可能抛出运行时异常,这证明当前客户端代码是不稳定的,需要记录异常,并排查解决问题。一旦有运行时异常,为了保证系统运行安全,也是抛出异常,由调用方注销局部会话,并且走重新登录流程.当然这个异常是系统自动抛出的。
+     *
+     * @param accessToken
+     */
+    private void authAccessToken(String accessToken) throws Exception{
+        try {
+            // 客户端登录校验通过的情况下,需要验证accessToken是否有效,目的有二:
+            // 1.当前accessToken如果无效,需要及时换取新的accessToken,以保持客户端和sso认证中心之间通信的畅通。
+            // 2.accessToken其实代表了sso全局会话,因为是一根绳,通过accessToken可以窥探全局会话是否失效,如果accessToken对应的sso全局会话失效,那么就要注销本地会话，跳转到登录页。
+            ResponseResult<Map> responseResult = ssoRemote.auth(accessToken);
+            logger.info(JSONObject.toJSONString(responseResult));
+
+            // 如果accessToken有效,则考虑其有效期剩余小于30秒的问题。
+            if (responseResult.getCode().equals("0")) {
+                // 取出remaining_expiration_time。
+                Map<String, String> map = responseResult.getData();
+                Object remainingExpirationTimeObj = map == null ? null : map.get("remaining_expiration_time");
+                String remainingExpirationTimeStr = remainingExpirationTimeObj == null ? null : String.valueOf(remainingExpirationTimeObj);
+                Long remainingExpirationTimeLong = remainingExpirationTimeStr == null ? null : Long.valueOf(remainingExpirationTimeStr);
+                if (null == map || null == remainingExpirationTimeLong) {
+                    throw new BusinessException(CoreConstants.REQUEST_PROGRAM_ERROR_CODE, "系统调用异常");
+                }
+
+                // 如果accessToken有效期剩余小于30秒,也认为过期,触发refresh_token换access_token过程。
+                // 有效期剩余小于30秒的accessToken是不安全的,继续使用它通信可能随时面临失效而引发失败的风险,所以也当作已过期处理。
+                if (remainingExpirationTimeLong < 30L ) {
+                    try {
+                        this.refreshToken(accessToken);
+                    } catch (Exception e) {
+                        // 一般情况是refresh_token也失效,导致其换取access_token失败,不管什么原因导致换取失败,此时都要抛出异常,由调用方注销局部会话,并且走重新登录流程.
+                        // 因为不注销也不行了,这根绳已经断了,不注销后续没有办法与sso用户中心通讯了,比如续期等操作都没办法做,只能注销通过重新登录重新把这根绳建立起来。
+                        logger.info(JSONObject.toJSONString(e));
+                        throw new BusinessException(CoreConstants.REQUEST_PROGRAM_ERROR_CODE, "refreshToken换取accessToken失败");
+                    }
+                }
+            // 如果accessToken失效,触发refresh_token换access_token过程。
+            } else if (responseResult.getCode().equals("50035")) {
+                try {
+                    this.refreshToken(accessToken);
+                } catch (Exception e) {
+                    // 一般情况是refresh_token也失效,导致其换取access_token失败,不管什么原因导致换取失败,此时都要抛出异常,由调用方注销局部会话,并且走重新登录流程.
+                    // 因为不注销也不行了,这根绳已经断了,不注销后续没有办法与sso用户中心通讯了,比如续期等操作都没办法做,只能注销通过重新登录重新把这根绳建立起来。
+                    logger.info(JSONObject.toJSONString(e));
+                    throw new BusinessException(CoreConstants.REQUEST_PROGRAM_ERROR_CODE, "refreshToken换取accessToken失败");
+                }
+            // 如果accessToken对应的sso全局会话失效,抛出异常,统一处理,即注销本地局部会话,走重新登录流程。
+            } else if (responseResult.getCode().equals("60032")) {
+                throw new BusinessException(CoreConstants.REQUEST_PROGRAM_ERROR_CODE, "sso全局会话失效");
+            // 其他异常情况,抛出异常,统一处理,即注销本地局部会话,走重新登录流程。
+            } else {
+                throw new BusinessException(CoreConstants.REQUEST_PROGRAM_ERROR_CODE, "sso全局会话通信异常");
+            }
+        } catch (Exception e) {
+            logger.info(JSONObject.toJSONString(e));
+            throw e;
+        }
+    }
+
+    /**
+     * refreshToken刷accessToken
+     * @param refreshToken
+     * @throws Exception
+     */
+    private void refreshToken(String refreshToken) throws Exception {
+        try {
+            // 使用refreshToken刷accessToken
+            OauthClientInfo oauthClientInfo = oauthClientInfoService.getOauthClientInfoByServerType("sso");
+            String client_id = oauthClientInfo.getClientId();
+
+            Map refreshTokenMap = ssoRemote.refreshToken(client_id, refreshToken, "refresh_token");
+            String access_token = null == refreshTokenMap ? null :(String)refreshTokenMap.get("access_token");
+            String expires_in = null == refreshTokenMap ? null : (String)refreshTokenMap.get("expires_in");
+            String refresh_token = null == refreshTokenMap ? null : (String)refreshTokenMap.get("refresh_token");
+//          String token_type = null == refreshTokenMap ? null : (String)refreshTokenMap.get("token_type");
+//          String scope = null == refreshTokenMap ? null : (String)refreshTokenMap.get("scope");
+//          Long userId = null == refreshTokenMap ? null : Long.valueOf((String)refreshTokenMap.get("userId"));
+
+            // 其他字段这里偷个懒不一一判断了
+            if (null == refreshTokenMap || StringUtils.isBlank(access_token)) {
+                throw new BusinessException("-1", "refreshToken换取accessToken失败");
+            }
+
+            // 更新accessToken信息
+            OauthAccessToken oauthAccessToken = new OauthAccessToken();
+            oauthAccessToken.setAccessToken(access_token);
+            oauthAccessToken.setExpireIn(StringUtils.isBlank(expires_in) ? null : Long.valueOf(expires_in));
+            oauthAccessToken.setRefreshToken(refresh_token);
+            oauthAccessTokenService.modifyById(oauthAccessToken);
+        } catch (Exception e) {
+            logger.info(JSONObject.toJSONString(e));
+            throw e;
+        }
     }
 
     @Override
